Refactor: IssueModificationFieldsDto -> IssueModificationFields로 변경
DTO는 계층 간 정보 전달에 사용하는 객체를 지칭하는 것이라 판단해 DTO 명칭을 제거하고, 해당 객체를 issue 패키지로 옮김.
