+++ BE/src/main/java/kr/codesquad/issuetraker/domain/issue/IssueModificationFields.java
@@ -1,4 +1,4 @@
-package kr.codesquad.issuetraker.dto;
+package kr.codesquad.issuetraker.domain.issue;
 
 import kr.codesquad.issuetraker.domain.label.Label;
 import kr.codesquad.issuetraker.domain.milestone.Milestone;
@@ -8,7 +8,7 @@ import lombok.Getter;
 
 @Builder
 @Getter
-public class IssueModificationFieldsDto {
+public class IssueModificationFields {
     private Milestone milestone;
     private Label label;
     private User author;
