--- BE/src/main/java/kr/codesquad/issuetraker/dto/IssueModificationType.java
@@ -1,10 +1,0 @@
-package kr.codesquad.issuetraker.dto;
-
-public enum IssueModificationType {
-    TITLE,
-    DESCRIPTION,
-    AUTHOR,
-    ASSIGNEE,
-    MILESTONE,
-    OPEN
-}
